TestTask2/requirements.txt
Django==4.1.4
django_crispy_forms==1.14.0
